# Add initial WASM SIMD128 support