--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,7 +1,10 @@
 [workspace]
 resolver = "2"
-members = ["fearless_simd", "fearless_simd_gen"]
+members = ["fearless_simd", "fearless_simd_gen", "fearless_simd_tests"]
 
 [workspace.package]
 license = "MIT/Apache-2.0"
 edition = "2024"
+
+[workspace.dependencies]
+fearless_simd = { path = "fearless_simd" }