--- a/fearless_simd/src/macros.rs
+++ b/fearless_simd/src/macros.rs
@@ -12,16 +12,24 @@ macro_rules! simd_dispatch {
     ) => {
         $( #[$meta] )* $vis
         fn $func(level: $crate::Level $(, $arg: $ty )*) $( -> $ret )? {
-            #[cfg(all(feature = "std", target_arch = "aarch64"))]
+            #[cfg(target_arch = "aarch64")]
             #[target_feature(enable = "neon")]
             #[inline]
             unsafe fn inner_neon(neon: $crate::aarch64::Neon $( , $arg: $ty )* ) $( -> $ret )? {
                 $inner( neon $( , $arg )* )
             }
+            #[cfg(target_arch = "wasm32")]
+            #[target_feature(enable = "simd128")]
+            #[inline]
+            unsafe fn inner_wasm_simd128(simd128: $crate::wasm32::WasmSimd128 $( , $arg: $ty )* ) $( -> $ret )? {
+                $inner( simd128 $( , $arg )* )
+            }
             match level {
                 Level::Fallback(fb) => $inner(fb $( , $arg )* ),
-                #[cfg(all(feature = "std", target_arch = "aarch64"))]
+                #[cfg(target_arch = "aarch64")]
                 Level::Neon(neon) => unsafe { inner_neon (neon $( , $arg )* ) }
+                #[cfg(target_arch = "wasm32")]
+                Level::WasmSimd128(wasm) => unsafe { inner_wasm_simd128 (wasm $( , $arg )* ) }
             }
         }
     };