--- a/fearless_simd_gen/src/arch/mod.rs
+++ b/fearless_simd_gen/src/arch/mod.rs
@@ -3,6 +3,7 @@
 
 pub(crate) mod fallback;
 pub(crate) mod neon;
+pub(crate) mod wasm;
 
 use proc_macro2::TokenStream;