--- a/fearless_simd_gen/src/main.rs
+++ b/fearless_simd_gen/src/main.rs
@@ -10,6 +10,7 @@ mod arch;
 mod generic;
 mod mk_fallback;
 mod mk_neon;
+mod mk_wasm;
 mod mk_ops;
 mod mk_simd_trait;
 mod mk_simd_types;
@@ -22,6 +23,7 @@ enum Module {
     SimdTrait,
     Ops,
     Neon,
+    Wasm,
     Fallback,
 }
 
@@ -38,6 +40,7 @@ impl Module {
             Module::SimdTrait => mk_simd_trait::mk_simd_trait(),
             Module::Ops => mk_ops::mk_ops(),
             Module::Neon => mk_neon::mk_neon_impl(mk_neon::Level::Neon),
+            Module::Wasm => mk_wasm::mk_wasm128_impl(mk_wasm::Level::WasmSimd128),
             Module::Fallback => mk_fallback::mk_fallback_impl(),
         }
     }
@@ -56,6 +59,7 @@ impl Module {
             Module::Ops => "ops",
             Module::Neon => "neon",
             Module::Fallback => "fallback",
+            Module::Wasm => "wasm",
         }
     }
 }