--- a/fearless_simd_gen/src/mk_fallback.rs
+++ b/fearless_simd_gen/src/mk_fallback.rs
@@ -32,13 +32,13 @@ pub fn mk_fallback_impl() -> TokenStream {
         use crate::{seal::Seal, Level, Simd, SimdInto};
 
         #imports
-        
+
         #[cfg(all(feature = "libm", not(feature = "std")))]
         trait FloatExt {
             fn floor(self) -> f32;
             fn sqrt(self) -> f32;
         }
-        
+
         #[cfg(all(feature = "libm", not(feature = "std")))]
         impl FloatExt for f32 {
             #[inline(always)]
@@ -50,7 +50,7 @@ pub fn mk_fallback_impl() -> TokenStream {
                 libm::sqrtf(self)
             }
         }
-        
+
         /// The SIMD token for the "fallback" level.
         #[derive(Clone, Copy, Debug)]
         pub struct Fallback {