> [!NOTE]
> The measurements in this PR were done very rough via a quick Chrome profile. Because the numbers are small, until we run a rigorous benchmark it'll be hard to validate the impact. Especially across browser engines, and with JS engines optimising WASM and fusing operators.

## Overview

This PR adds initial WASM SIMD support to `fearless_simd`, implementing enough operations to enable WASM SIMD in https://github.com/linebender/vello/pull/1053. Rather than implementing all operations in one large PR, this focuses on the essential subset needed for Vello and breaks up an otherwise huge change. There's also some tricky operations to implement using the small amount of WASM instructions 😅 .

## Performance Impact

Tested with Ghost Tiger rendering in Vello:

### Without fast kurbo (baseline):
- **Without SIMD**: ~42.31ms per frame
- **With SIMD**: ~39.91ms per frame  
- **Improvement**: ~5% faster

### With fast kurbo (https://github.com/linebender/kurbo/pull/427):
- **Without SIMD**: ~6ms per frame
- **With SIMD**: ~4ms per frame
- **Improvement**: ~30% faster (Take this with a huge grain of salt)

*Test methodology: I linked `vello` locally to `fearless_simd` via `path` reference, and modified `vello_hybrid` to use the WASM SIMD level.*

## Changes

- **New architecture**: Added WASM SIMD128 support
- **Operations implemented**: Core subset including:
  - Binary ops: `add`, `sub`, `mul`, `min`, `max`
  - Comparison ops: `simd_eq`, `simd_ne`, `simd_lt`, etc.
  - Math ops: `sqrt`, `madd`
- **Testing**: Added parity tests ensuring Fallback and WASM SIMD produce identical results
- **Bug fix**: Fixed incorrect mask generation in Fallback comparison operations (was returning `0/1` instead of `0/-1`)

## Test Plan

Added `test_wasm_simd_parity!` macro that verifies operations produce identical results across Fallback and WASM SIMD implementations.

I only tested a small subset. Maybe in the future we code-gen the tests as well?

## Next Steps

Future PRs will add more operations to achieve full WASM SIMD coverage.